+++ R/99-utils.R
@@ -1,17 +1,47 @@
-#' Create a project directory to use with 'KuskoHarvEst'
+#' Create path to a package resource file
+#'
+#' Constructs a complete file path to a package resource file located in
+#'   `rstudio/templates/project/resources` within the 'KuskoHarvEst' library
+#'
+#' @param file Character; a file name or file path within the resources folder to point to
+#'
+
+resource_path = function(file) {
+  system.file(file.path("rstudio", "templates", "project", "resources", file), package = "KuskoHarvEst")
+}
+
+#' Create a project directory to use for use with 'KuskoHarvEst'
 #'
 #' Called by the RStudio project template builder
 #'
 #' @param path A location to put the new project
+#' @param is_for_final_report Logical; is the project for compiling all estimates
+#'   into tables and figures for final reporting rather than producing estimates for one day in-season?
 #'
 
-KuskoHarvEst_skeleton = function(path) {
+KuskoHarvEst_skeleton = function(path, is_for_final_report) {
 
-  # create the package directory
+  # create the project directory
   dir.create(path, recursive = TRUE, showWarnings = FALSE)
 
-  # create subdirectories
-  dir.create(file.path(path, "data-raw"))
+  if (!is_for_final_report) {
+    # create subdirectories
+    dir.create(file.path(path, "data-raw"))
+  } else {
+    # create subdirectories
+    dir.create(file.path(path, "raw-data-files"))
+
+    # find the files
+    resource_path = resource_path("07-final-report-content")
+    files = list.files(resource_path)
+
+    # build full file paths
+    source = file.path(resource_path, files)
+    target = file.path(path, files)
+
+    # copy the files into the new project
+    file.copy(source, target)
+  }
 
   TRUE
 }
@@ -33,25 +63,6 @@ combine_datetime = function(dates, times) {
 
   # return the output
   return(step2)
-}
-
-#' Combine a list of data frames
-#'
-#' Loops over elements of a list object, where each element is a data frame with identical headers
-#'   and applies [base::rbind()] to them.
-#'
-#' @param list List; each element is a data frame with identical headers
-#'
-
-unlist_dfs = function(list) {
-  # empty object
-  output = NULL
-
-  # loop through list elements, combining the data frame in each with all previous
-  for (i in 1:length(list)) output = rbind(output, list[[i]])
-
-  # return the output
-  return(output)
 }
 
 #' Convert a proportion to a percent
@@ -208,4 +219,22 @@ add_vspace = function(kable_input, space = "-1em") {
 
 link_to_doc = function(doc, text = "here") {
   paste0('[', text, '](./', doc, '){target="_blank"}')
+}
+
+#' Summation-informed rounding
+#'
+#' Rounds a vector such that the sum of the rounded vector equals the sum of the unrounded vector
+#'
+#' @param x Numeric; vector to be rounded
+#' @param digits Numeric; number of decimal points to round to
+#' @references The source code for this function was copied from [this Stack Overflow answer](https://stackoverflow.com/a/35930285/3911200)
+
+smart_round = function(x, digits = 0) {
+  # copied from https://stackoverflow.com/a/35930285/3911200
+  up = 10 ^ digits
+  x = x * up
+  y = floor(x)
+  indices = tail(order(x-y), round(sum(x)) - sum(y))
+  y[indices] = y[indices] + 1
+  y/up
 }